Desktop/onCampus/backend/requirements.txt
asgiref==3.6.0
boto3==1.26.55
botocore==1.29.55
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
dj-rest-auth[with_social]==2.2.5
markdown==3.4.1
future==0.18.3
geocoder==1.38.1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
ratelim==0.1.6
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0
